+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## UNRELEASED
+- Feature: Enable `producers.persistent_connections` phobos setting
 
 # 1.24.2 - 2024-05-01
 - Fix: Deprecation notice with Rails 7.